Moved existing classes to a more descriptive package.
