+++ src/test/java/com/tinkerpop/furnace/algorithms/DegreeCentralityTest.java
@@ -4,6 +4,7 @@ import com.tinkerpop.blueprints.Direction;
 import com.tinkerpop.blueprints.Graph;
 import com.tinkerpop.blueprints.Vertex;
 import com.tinkerpop.blueprints.impls.tg.TinkerGraphFactory;
+import com.tinkerpop.furnace.algorithms.rank.DegreeCentrality;
 import com.tinkerpop.furnace.util.QueryTemplate;
 import junit.framework.TestCase;
